experimentoConfidencialidad/requirements.txt
aws-cdk-lib==2.213.0
constructs>=10.0.0,<11.0.0
boto3>=1.26.0
requests>=2.28.0
PyJWT>=2.8.0
cryptography>=3.4.8
pytz>=2023.3
psycopg2-binary>=2.9.0
ipaddress>=1.0.23
